Generalize public_* checks into noisy_* checks
We had public_command_authed, public_check_priv and
public_check_antipriv to verify that someone was authorized and complain
loudly (into the configured channel) otherwise. In preparation for
supporting multiple channels, and also in order to make everything more
sensible, this renames the functions and requires that a return path is
passed to them. This way, callers can decide where to send the
complaint. It might often be preferable to not annoy the channel with
random complaints about privileges, after all.

Signed-off-by: Jan Krüger <jk@jk.gs>