Introduce new response API for handlers
This is a big set of changes. It allows handlers (command handlers and
IRC event handlers) to automatically send responses to the right places.
It can also automatically redirect output to a different channel (if
desired) and will, by default, include an address in a public response
if one was provided in a request.

All of this is implemented by keeping an internal handler context
structures. It is initialised with event parameters (sender of message,
relevant targets, ...) before each handler is called, and cleared
afterwards. Handlers call BotIrc::check_ctx() to

    (a)  signal that they want to use the response API and
    (b)  configure response behaviour and privileges.

check_ctx can be called multiple times within a handler to check
privileges as needed.

The response API distinguishes between "noise" responses (information
related to running the command, e.g. status/authorization messages) and
"wisdom" responses (the actual information payload requested as part of
the handled event). Automatic addressing and redirection only apply to
wisdom messages since event results are really only interesting to the
source of the event and not random other people/places.

Signed-off-by: Jan Krüger <jk@jk.gs>