BotIrc::return_path: fortify against weird arguments
Since prepare_handler_ctx will not always have an arrayref of targets
available but return_path (that is called by it) expects one,
dereferencing there will sometimes kill the bot. This is, of course,
unacceptable. Ignoring the targets is the right thing to do because we
still have the source of the event as a potential return path.

Signed-off-by: Jan Krüger <jk@jk.gs>