plugins/voice: ctl command: fix incorrect code sending confirmation
Whoops. Originally I was going to use the standard BotCtl::send
interface for the confirmation, but halfway through I changed my mind
because it wasn't really necessary for sending a mere "ok". I only
changed half of the line, though. This gets rid of the superfluous
argument that resulted from my being sloppy.

Signed-off-by: Jan Krüger <jk@jk.gs>