BotIrc::ctx_set_addressee: make regex easier to read and fix bug
This simplifies the regex for matching nick prefixes in public messages
and also fixes the issue that it wouldn't match one-character nicknames.
Since we recently added a check that the nickname actually exists in the
channel, we can also afford to make the nick part of the regex less
strict.

Signed-off-by: Jan Krüger <jk@jk.gs>
Secret-nod-to: Eugene E. Kashpureff Jr <eugene@kashpureff.org>